C# Inheritance Without Constructors
In this lesson we look at inheritance.
Inheritance is a fundamental concept in object-oriented programming that allows a child class to inherit the properties from the parent class. 
The new class inherits the properties and methods of the existing class and can also add new properties and methods of its own.